sköll: Add material parsing from assimp data
This adds material paring from the assimp data we get. This will
eventually just be replaced by a custom compact data format, but for now
we'll continue to parse with assimp.

Next I need to add defaults so that if you don't have a texture, we at
least do something other than just display black